Update to 24.12.12

+++ AnyCAD.Advanced/Simulation/Simulation_RayHit.cs
@@ -28,7 +28,7 @@ namespace AnyCAD.Demo.Graphics
             camera.UpdateProjectionMatrix();
 
             var ray = new Ray(new Vector3d(0, 100, 0), -Vector3d.UNIT_Y);
-            mCaster = new Raycaster(camera, (uint)EnumShapeFilter.Face, ray, 50,50);
+            mCaster = new Raycaster(camera, EnumShapeFilter.Face, ray, 50,50);
 
             mLine = new SegmentsSceneNode(1, ColorTable.Red, 1);
             render.ShowSceneNode(mLine);
